Separate out infrastructure metrics, include in predefined metrics sets
